Fix two flaky tests in the connectors test assembly
WhenAQueuedWriteIsOverwrittenByInitialState_ThenTheLocalWriteStillWins fenced its
assertion behind a write to a second property. The retry queue reconcile sends a change
the model already holds straight to the source, while a change the initial-state load
moved the model off is restored locally and left to the connected processor, which
delivers it a buffer time later. The fence therefore arrived first, and the stop that
followed cancelled the processor before it flushed. It now waits for the write it
asserts on, which is also the pattern the neighbouring tests already use.

The snapshots under Updates assume two sequential writes carry distinct timestamps,
because Verify names scrubbed timestamps by value. DateTime.UtcNow resolves to a
microsecond, which two property writes stay well inside once the process is warm, so
LastName rendered DateTimeOffset_1 where the snapshot records DateTimeOffset_2. That
made the failure appear only in a full-assembly run and never in isolation. The tracking
test assembly already installed a monotonic clock for the same reason, so that clock
moves to the shared testing library and each assembly keeps only the module initializer
that installs it, which has to be per assembly.

Both are test defects. Writes that share a timestamp are correct: nothing in the
delivery path orders by timestamp, which is what the commit revision is for.
